deps: upgrade v8 to 4.5.92
This commit includes two fix-ups to src/node_contextify.cc and
lib/module.js to force-load the debugger when necessary.

PR-URL: https://github.com/nodejs/io.js/pull/2091

+++ deps/v8/src/arm64/macro-assembler-arm64.h
@@ -886,8 +886,8 @@ class MacroAssembler : public Assembler {
 
   template<typename Field>
   void DecodeField(Register dst, Register src) {
-    static const uint64_t shift = Field::kShift;
-    static const uint64_t setbits = CountSetBits(Field::kMask, 32);
+    static const int shift = Field::kShift;
+    static const int setbits = CountSetBits(Field::kMask, 32);
     Ubfx(dst, src, shift, setbits);
   }
